Split up user token files
