+++ Sources/App/Modules/User/Database/Models/UserToken.swift
@@ -3,37 +3,6 @@ import Fluent
 import Vapor
 import JWT
 
-// TODO: split files, add migration
-
-final class UserTokenFamilyModel: DatabaseModelInterface, Authenticatable {
-    typealias Module = UserModule
-
-    enum FieldKeys {
-        enum v1 {
-            static var tokenType: FieldKey { "token_type" }
-            static var lastTokenRefresh: FieldKey { "last_token_refresh" }
-            static var userId: FieldKey { "user_id" }
-        }
-    }
-
-    @ID() var id: UUID?
-    @Enum(key: FieldKeys.v1.tokenType) var tokenType: UserTokenType
-    @Field(key: FieldKeys.v1.lastTokenRefresh) var lastTokenRefresh: Date
-    @Parent(key: FieldKeys.v1.userId) var user: UserAccountModel
-
-    init() { }
-
-    init(
-        id: UUID? = nil,
-        userId: UUID,
-        tokenType: UserTokenType
-    ) {
-        self.id = id
-        self.$user.id = userId
-        self.tokenType = tokenType
-    }
-}
-
 struct UserToken: JWTPayload, Equatable {
     /// Calculates the expiration date for a given token type.
     /// - Parameter tokenType: The type of token for which the expiration date should be calculated.
@@ -204,19 +173,4 @@ extension UserToken {
             tokenFamily: tokenFamily.requireID()
         )
     }
-}
-
-enum UserTokenType: String, Codable, DatabaseEnumInterface {
-    typealias Module = UserModule
-
-    static let schema = "user_token_type"
-    
-    /// Token used to access content.
-    case contentAccess
-
-    /// Token used to refresh access tokens.
-    case tokenRefresh
-
-    /// Token used to verify a user's identity. Usually sent via E-Mail..
-    case verification
 }
